requirements.txt
# CABiNet - Semantic Segmentation Requirements

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.7.0

# Configuration Management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Visualization & Logging
tqdm>=4.65.0
rich>=13.3.0

# Development Tools
pytest>=7.3.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.3.0

# Optional: Distributed Training
# torch.distributed (included with PyTorch)
